Rename files in Partners module
