+++ src/test/resources/features/Partners/Partners_TC02.feature
@@ -1,6 +1,6 @@
 Feature: Create a new corporate customer partner
 
-  @TC_02
+  @Partners_TC02
   Scenario: Create new partner, who is corporate and customer
     Given Navigate to the partners module and click the all partners
     Given Create new partner
